Fix controller naming and related issues
